+++ firmware/config/boards/hellen/hellenNA8_96/board_configuration.cpp
@@ -65,7 +65,6 @@ static void setupDefaultSensorInputs() {
 	// trigger inputs, hall
 	engineConfiguration->triggerInputPins[0] = H144_IN_CAM;
 	engineConfiguration->triggerInputPins[1] = H144_IN_CRANK;
-	engineConfiguration->triggerInputPins[2] = Gpio::Unassigned;
 	engineConfiguration->camInputs[0] = Gpio::Unassigned;
 
 	engineConfiguration->tps1_1AdcChannel = EFI_ADC_4;